Factor out AdminController into UserController and RefreshTokenController
