+++ src/Controller/UserController.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\Response\EmptyBody;
 use App\Response\ErrorResponse;
 use App\Response\ErrorResponseBody;
 use App\Response\LabelledBody;
@@ -12,7 +11,6 @@ use App\Response\Response;
 use App\Response\User\User;
 use App\Security\AuthenticationToken;
 use App\Security\UserCredentials;
-use App\Security\UserId;
 use Psr\Http\Client\ClientExceptionInterface;
 use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
 use SmartAssert\ServiceClient\Exception\InvalidResponseDataException;
@@ -24,16 +22,15 @@ use SmartAssert\UsersClient\Exception\UserAlreadyExistsException;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/admin', name: 'admin_')]
-readonly class AdminController
+readonly class UserController
 {
     public function __construct(
         private Client $client
     ) {
     }
 
-    #[Route('/user/create', name: 'user_create', methods: ['POST'])]
-    public function createUser(AuthenticationToken $token, UserCredentials $userCredentials): JsonResponse
+    #[Route('/admin/user/create', name: 'admin_user_create', methods: ['POST'])]
+    public function create(AuthenticationToken $token, UserCredentials $userCredentials): JsonResponse
     {
         try {
             $user = $this->client->createUser(
@@ -126,54 +123,6 @@ readonly class AdminController
                 'user',
                 new User($user->id, $user->userIdentifier)
             )
-        );
-    }
-
-    #[Route('/revoke-refresh-token', name: 'revoke_refresh_token', methods: ['POST'])]
-    public function revokeRefreshToken(AuthenticationToken $token, UserId $userId): JsonResponse
-    {
-        try {
-            $this->client->revokeFrontendRefreshToken($token->token, $userId->id);
-        } catch (ClientExceptionInterface $e) {
-            $code = $e->getCode();
-            $message = $e->getMessage();
-
-            return new ErrorResponse(
-                new ErrorResponseBody(
-                    'service-communication-failure',
-                    [
-                        'service' => 'users',
-                        'error' => [
-                            'code' => $code,
-                            'message' => $message,
-                        ],
-                    ]
-                )
-            );
-        } catch (UnauthorizedException) {
-            return new ErrorResponse(new ErrorResponseBody('unauthorized'), 401);
-        } catch (NonSuccessResponseException $e) {
-            if (404 === $e->getStatusCode()) {
-                return new ErrorResponse(
-                    new ErrorResponseBody('not-found'),
-                    $e->getStatusCode()
-                );
-            }
-
-            return new ErrorResponse(
-                new ErrorResponseBody(
-                    'non-successful-service-response',
-                    [
-                        'service' => 'users',
-                        'status' => $e->getStatusCode(),
-                        'message' => $e->getMessage(),
-                    ]
-                )
-            );
-        }
-
-        return new Response(
-            new EmptyBody()
         );
     }
 }
