+++ tests/Unit/Controller/User/UserControllerTest.php
@@ -4,11 +4,10 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\Controller\User;
 
-use App\Controller\AdminController;
+use App\Controller\UserController;
 use App\Response\ErrorResponse;
 use App\Security\AuthenticationToken;
 use App\Security\UserCredentials;
-use App\Security\UserId;
 use GuzzleHttp\Exception\TransferException;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\ClientExceptionInterface;
@@ -25,16 +24,15 @@ use SmartAssert\ServiceClient\Response\Response as ServiceClientResponse;
 use SmartAssert\UsersClient\Client;
 use SmartAssert\UsersClient\Model\RefreshableToken as UsersClientRefreshableToken;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Response;
 
-class AdminControllerTest extends TestCase
+class UserControllerTest extends TestCase
 {
     /**
-     * @dataProvider createUserUsersClientExceptionDataProvider
+     * @dataProvider createUsersClientExceptionDataProvider
      *
      * @param array<mixed> $expectedResponseData
      */
-    public function testCreateUser(
+    public function testCreate(
         \Exception $exception,
         int $expectedResponseStatusCode,
         array $expectedResponseData,
@@ -52,16 +50,23 @@ class AdminControllerTest extends TestCase
             ->andThrow($exception)
         ;
 
-        $controller = new AdminController($client);
-        $response = $controller->createUser($authenticationToken, $userCredentials);
+        $controller = new UserController($client);
+        $response = $controller->create($authenticationToken, $userCredentials);
 
-        $this->assertResponse($response, $expectedResponseStatusCode, $expectedResponseData);
+        self::assertSame($expectedResponseStatusCode, $response->getStatusCode());
+        self::assertInstanceOf(ErrorResponse::class, $response);
+        self::assertInstanceOf(JsonResponse::class, $response);
+        self::assertSame('application/json', $response->headers->get('content-type'));
+
+        $responseData = json_decode((string) $response->getContent(), true);
+
+        self::assertEquals($expectedResponseData, $responseData);
     }
 
     /**
      * @return array<mixed>
      */
-    public function createUserUsersClientExceptionDataProvider(): array
+    public function createUsersClientExceptionDataProvider(): array
     {
         $exceptionMessage = md5((string) rand());
         $exceptionCode = rand();
@@ -238,144 +243,5 @@ class AdminControllerTest extends TestCase
                 ],
             ],
         ];
-    }
-
-    /**
-     * @dataProvider revokeRefreshTokenUsersClientExceptionDataProvider
-     *
-     * @param array<mixed> $expectedResponseData
-     */
-    public function testRevokeRefreshToken(
-        \Exception $exception,
-        int $expectedResponseStatusCode,
-        array $expectedResponseData,
-    ): void {
-        $token = md5((string) rand());
-        $authenticationToken = new AuthenticationToken($token);
-        $id = md5((string) rand());
-        $userId = new UserId($id);
-
-        $client = \Mockery::mock(Client::class);
-        $client
-            ->shouldReceive('revokeFrontendRefreshToken')
-            ->with($token, $id)
-            ->andThrow($exception)
-        ;
-
-        $controller = new AdminController($client);
-        $response = $controller->revokeRefreshToken($authenticationToken, $userId);
-
-        $this->assertResponse($response, $expectedResponseStatusCode, $expectedResponseData);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function revokeRefreshTokenUsersClientExceptionDataProvider(): array
-    {
-        $exceptionMessage = md5((string) rand());
-        $exceptionCode = rand();
-
-        return [
-            ClientExceptionInterface::class => [
-                'exception' => new TransferException(
-                    $exceptionMessage,
-                    $exceptionCode
-                ),
-                'expectedResponseStatusCode' => 500,
-                'expectedResponseData' => [
-                    'type' => 'service-communication-failure',
-                    'context' => [
-                        'service' => 'users',
-                        'error' => [
-                            'code' => $exceptionCode,
-                            'message' => $exceptionMessage,
-                        ],
-                    ],
-                ],
-            ],
-            CurlExceptionInterface::class => [
-                'exception' => new CurlException(
-                    \Mockery::mock(RequestInterface::class),
-                    $exceptionCode,
-                    $exceptionMessage,
-                ),
-                'expectedResponseStatusCode' => 500,
-                'expectedResponseData' => [
-                    'type' => 'service-communication-failure',
-                    'context' => [
-                        'service' => 'users',
-                        'error' => [
-                            'code' => $exceptionCode,
-                            'message' => $exceptionMessage,
-                        ],
-                    ],
-                ],
-            ],
-            NonSuccessResponseException::class . ' 404' => [
-                'exception' => new NonSuccessResponseException(
-                    (function () {
-                        $response = \Mockery::mock(ResponseInterface::class);
-                        $response
-                            ->shouldReceive('getStatusCode')
-                            ->andReturn(404)
-                        ;
-
-                        $response
-                            ->shouldReceive('getReasonPhrase')
-                            ->andReturn('Not found.')
-                        ;
-
-                        return $response;
-                    })(),
-                ),
-                'expectedResponseStatusCode' => 404,
-                'expectedResponseData' => [
-                    'type' => 'not-found',
-                ],
-            ],
-            NonSuccessResponseException::class . ' 405' => [
-                'exception' => new NonSuccessResponseException(
-                    (function () {
-                        $response = \Mockery::mock(ResponseInterface::class);
-                        $response
-                            ->shouldReceive('getStatusCode')
-                            ->andReturn(405)
-                        ;
-
-                        $response
-                            ->shouldReceive('getReasonPhrase')
-                            ->andReturn('Method not allowed.')
-                        ;
-
-                        return $response;
-                    })(),
-                ),
-                'expectedResponseStatusCode' => 500,
-                'expectedResponseData' => [
-                    'type' => 'non-successful-service-response',
-                    'context' => [
-                        'service' => 'users',
-                        'status' => 405,
-                        'message' => '405: Method not allowed.',
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * @param array<mixed> $expectedData
-     */
-    private function assertResponse(Response $response, int $expectedCode, array $expectedData): void
-    {
-        self::assertSame($expectedCode, $response->getStatusCode());
-        self::assertInstanceOf(ErrorResponse::class, $response);
-        self::assertInstanceOf(JsonResponse::class, $response);
-        self::assertSame('application/json', $response->headers->get('content-type'));
-
-        $responseData = json_decode((string) $response->getContent(), true);
-
-        self::assertEquals($expectedData, $responseData);
     }
 }
